SMOKEZIP: add Author, Revision and Date keywords to smokezip source
Former-commit-id: 78c4602635de2dc12bc5513fbbc551075c77fc6c [formerly c3f20ffb78f1bd89b8f02d440102a70979bc7836]
Former-commit-id: 56fc275b839aacb5add21fab0b2cb218687cc321
Former-commit-id: 70d14c7c7d07bb3528b84428a57a7b838b31cb96

+++ SMV_5/source/smokezip/CNV3dsmoke.c
@@ -1,3 +1,7 @@
+// $Date$ 
+// $Revision$
+// $Author$
+
 #include "options.h"
 #include <stdio.h>
 #include <stdlib.h>
@@ -355,7 +359,7 @@ void convert_3dsmoke(smoke3d *smoke3di){
     if(make_lighting_file==1)fprintf(smoke3dsizestream," %i",smoke3di->ncompressed_lighting_zlib);
     fprintf(smoke3dsizestream,"\n");
 #else
-    fprintf(smoke3dsizestream,"%f %i %i %i\n",time,nfull,ncompressed_rle,ncompressed_zlib);
+    fprintf(smoke3dsizestream,"%f %i %i %i\n",time,nfull,ncompressed_rle,(int)ncompressed_zlib);
 #endif
   }
   printf(" 100%s completed\n",pp);
